update license script

+++ packages/vscode-extension/build/copy_license.js
@@ -4,4 +4,7 @@ const path = require("path");
 const licensePath = path.join(__dirname, "../../../LICENSE");
 const targetPath = path.join(__dirname, "../LICENSE");
 
+if(fs.existsSync(targetPath)) {
+  return;
+}
 fs.copyFileSync(licensePath, targetPath);
